ComputeAverage now uses a double to account for large number of samples

+++ devices/Hx711/Hx711.cs
@@ -214,7 +214,7 @@ namespace Iot.Device.Hx711
 
         private int ComputeAverage(int[] values)
         {
-            long value = 0;
+            double value = 0;
 
             for (int i = 0; i < values.Length; i++)
             {
